refactor(api): move duplicate setter code to mixin
